requirements.txt
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2==2.9.5
django-filter==22.1
djangorestframework==3.14.0
environ~=1.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0


